separate event listener for example producer
