+++ wrappers/python/wpwithin_python/examples/producer_callbacks.py
@@ -1,21 +1,7 @@
 import signal
 import sys
-from wpwithin_python import create_client, PricePerUnit, Price, Service, AbstractEventListener
-
-
-class EventListener(AbstractEventListener):
-    def begin_service_delivery(self, service_id, service_delivery_token, units_to_supply):
-        print("Begin Service Delivery")
-        print("Service ID: {0}".format(service_id))
-        print("Units to supply: {0}".format(units_to_supply))
-        print("Service delivery token: {0}".format(service_delivery_token))
-
-    def end_service_delivery(self, service_id, service_delivery_token, units_received):
-        print("End Service Delivery")
-        print("Service ID: {0}".format(service_id))
-        print("Units to supply: {0}".format(units_received))
-        print("Service delivery token: {0}".format(service_delivery_token))
-
+from wpwithin_python import create_client, PricePerUnit, Price, Service
+from callbacks_event_listener import EventListener
 
 out = create_client("127.0.0.1",
                     9090,
